+++ builder/upcloud/artifact_test.go
@@ -6,6 +6,8 @@ import (
 
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
+	"github.com/hashicorp/packer-plugin-sdk/packer/registry/image"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestArtifact_impl(t *testing.T) {
@@ -41,4 +43,50 @@ func TestArtifact_String(t *testing.T) {
 	if result != expected {
 		t.Errorf("Expected: %q, got: %q", expected, result)
 	}
+}
+
+func TestArtifact_Metadata(t *testing.T) {
+	templates := []*upcloud.Storage{}
+	templates = append(templates,
+		&upcloud.Storage{
+			UUID:  "some-uuid",
+			Size:  10,
+			Title: "some-title",
+			Zone:  "fi-hel1",
+		},
+		&upcloud.Storage{
+			UUID:  "some-other-uuid",
+			Size:  10,
+			Title: "some-title",
+			Zone:  "fi-hel2",
+		},
+	)
+
+	a := &Artifact{
+		Templates: templates,
+		config: &Config{
+			Zone:           "fi-hel1",
+			CloneZones:     []string{"fi-hel2"},
+			TemplatePrefix: "prefix",
+		},
+		StateData: map[string]interface{}{
+			"source_template_title": "source-title",
+			"source_template_uuid":  "source-uuid",
+		},
+	}
+	got := a.State(image.ArtifactStateURI).([]*image.Image)
+	want := &image.Image{
+		ImageID:        "some-uuid",
+		ProviderName:   "upcloud",
+		ProviderRegion: "fi-hel1",
+		Labels: map[string]string{
+			"source":      "source-title",
+			"source_id":   "source-uuid",
+			"name":        "some-title",
+			"name_prefix": "prefix",
+			"size":        "10",
+		},
+		SourceImageID: "source-uuid",
+	}
+	assert.Equal(t, want, got[0])
 }